consolidate signout
